Generate serializers for class LayoutRect, ScreenDataOverrides, MediaCapabilitiesInfo, MediaStreamRequest https://bugs.webkit.org/show_bug.cgi?id=247648 rdar://101308191
Reviewed by Alex Christensen.

* Source/WebCore/platform/MediaCapabilitiesInfo.h:
(WebCore::MediaCapabilitiesInfo::encode const): Deleted.
(WebCore::MediaCapabilitiesInfo::decode): Deleted.
* Source/WebCore/platform/graphics/LayoutRect.h:
(WebCore::LayoutRect::encode const): Deleted.
(WebCore::LayoutRect::decode): Deleted.
* Source/WebCore/platform/graphics/VP9Utilities.h:
(WebCore::ScreenDataOverrides::encode const): Deleted.
(WebCore::ScreenDataOverrides::decode): Deleted.
* Source/WebCore/platform/mediastream/MediaStreamRequest.h:
(WebCore::MediaStreamRequest::encode const): Deleted.
(WebCore::MediaStreamRequest::decode): Deleted.
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:

Canonical link: https://commits.webkit.org/256545@main
