+++ Source/WebCore/platform/graphics/LayoutRect.h
@@ -34,6 +34,7 @@
 #include "IntRect.h"
 #include "LayoutPoint.h"
 #include "LengthBox.h"
+#include <wtf/ArgumentCoder.h>
 #include <wtf/Forward.h>
 
 namespace WTF {
@@ -57,28 +58,6 @@ public:
     LayoutRect(const IntRect& rect) : m_location(rect.location()), m_size(rect.size()) { }
     
     WEBCORE_EXPORT explicit LayoutRect(const FloatRect&); // don't do this implicitly since it's lossy
-
-    template<class Encoder>
-    void encode(Encoder& encoder) const
-    {
-        encoder << m_location << m_size;
-    }
-
-    template<class Decoder>
-    static std::optional<LayoutRect> decode(Decoder& decoder)
-    {
-        std::optional<LayoutPoint> layoutPoint;
-        decoder >> layoutPoint;
-        if (!layoutPoint)
-            return std::nullopt;
-
-        std::optional<LayoutSize> layoutSize;
-        decoder >> layoutSize;
-        if (!layoutSize)
-            return std::nullopt;
-
-        return {{ *layoutPoint, *layoutSize }};
-    }
 
     LayoutPoint location() const { return m_location; }
     LayoutSize size() const { return m_size; }
@@ -224,6 +203,7 @@ public:
     operator FloatRect() const { return FloatRect(m_location, m_size); }
 
 private:
+    friend struct IPC::ArgumentCoder<WebCore::LayoutRect, void>;
     void setLocationAndSizeFromEdges(LayoutUnit left, LayoutUnit top, LayoutUnit right, LayoutUnit bottom);
 
     LayoutPoint m_location;
