+++ Source/WebCore/platform/mediastream/MediaStreamRequest.h
@@ -35,35 +35,16 @@
 namespace WebCore {
 
 struct MediaStreamRequest {
-    enum class Type { UserMedia, DisplayMedia, DisplayMediaWithAudio };
+    enum class Type : uint8_t {
+        UserMedia,
+        DisplayMedia,
+        DisplayMediaWithAudio
+    };
     Type type { Type::UserMedia };
     MediaConstraints audioConstraints;
     MediaConstraints videoConstraints;
     bool isUserGesturePriviledged { false };
     PageIdentifier pageIdentifier;
-
-    template<class Encoder>
-    void encode(Encoder& encoder) const
-    {
-        encoder << type;
-        encoder << audioConstraints;
-        encoder << videoConstraints;
-        encoder << isUserGesturePriviledged;
-        encoder << pageIdentifier;
-    }
-
-    template <class Decoder> static std::optional<MediaStreamRequest> decode(Decoder& decoder)
-    {
-        MediaStreamRequest request;
-        if (!decoder.decode(request.type)
-            || !decoder.decode(request.audioConstraints)
-            || !decoder.decode(request.videoConstraints)
-            || !decoder.decode(request.isUserGesturePriviledged)
-            || !decoder.decode(request.pageIdentifier))
-            return std::nullopt;
-
-        return request;
-    }
 };
 
 } // namespace WebCore
@@ -73,23 +54,14 @@ struct MediaStreamRequest {
 namespace WebCore {
 
 struct MediaStreamRequest {
-    enum class Type { UserMedia, DisplayMedia, DisplayMediaWithAudio };
+    enum class Type : uint8_t {
+        UserMedia,
+        DisplayMedia,
+        DisplayMediaWithAudio
+    };
     Type type;
 };
 
 } // namespace WebCore
 
 #endif // ENABLE(MEDIA_STREAM)
-
-namespace WTF {
-
-template<> struct EnumTraits<WebCore::MediaStreamRequest::Type> {
-    using values = EnumValues<
-        WebCore::MediaStreamRequest::Type,
-        WebCore::MediaStreamRequest::Type::UserMedia,
-        WebCore::MediaStreamRequest::Type::DisplayMedia,
-        WebCore::MediaStreamRequest::Type::DisplayMediaWithAudio
-    >;
-};
-
-} // namespace WTF
